+++ gen3_util/config/cli.py
@@ -2,9 +2,9 @@ import subprocess
 
 import click
 
+from gen3_util import Config
 from gen3_util.cli import NaturalOrderGroup, CLIOutput
 from gen3_util.config import ensure_auth
-from gen3_util.config.config import Config
 
 
 @click.group(name='config', cls=NaturalOrderGroup)
