+++ gen3_util/files/manifest.py
@@ -18,7 +18,7 @@ import sys
 from gen3_util.buckets import get_program_bucket
 from gen3_util.config import Config, gen3_services
 from gen3_util.files.uploader import _normalize_file_url
-from gen3_util.meta import ACED_NAMESPACE
+from gen3_util import ACED_NAMESPACE
 from gen3_util.meta.importer import md5sum
 from tqdm import tqdm
